requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
neo4j==5.14.1
python-dotenv==1.0.0
langchain==0.0.335
langchain-core==0.1.17
langgraph==0.0.21
openai==1.3.7
celery==5.3.4
redis==5.0.1
python-multipart==0.0.6
pyjwt==2.8.0
pytest==7.4.3
tree-sitter==0.20.1
pandas==2.1.3
websockets==12.0
zipfile36==0.1.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1 